% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bskernel-package.R
\docType{package}
\name{bskernel-package}
\alias{bskernel}
\alias{bskernel-package}
\title{bskernel: Flexible Auto-covariance Construction via Spline Kernels}
\description{
This package provides flexible auto-covariance kernel construction via inverse Fourier transforms of B-spline bases, as well as some handy tools for optimisation. It accompanies the paper ''Universal approximation of auto-covariance functions via spline kernels''.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://astfalckl.github.io/bskernel/}
}

}
\author{
\strong{Maintainer}: Lachlan Astfalck \email{lachlan.astfalck@uwa.edu.au} (\href{https://orcid.org/0000-0002-8369-9384}{ORCID})

}
\keyword{internal}
