% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline_normalisation_constant}
\alias{bspline_normalisation_constant}
\title{Normalisation Constant for Truncated Power Expansion of B-spline}
\usage{
bspline_normalisation_constant(knots, i, k)
}
\arguments{
\item{knots}{A numeric vector of knots.}

\item{i}{An integer index indicating the position in the knot sequence for which to compute the coefficients.}

\item{k}{The degree of the B-spline basis function (must be 0, 1, or 2).}
}
\value{
A numeric scalar giving the inverse of the integral under the truncated power representation - i.e., a
multiplicative constant \eqn{Z^{-1}} that makes the basis function integrate to 1.
}
\description{
Computes the scaling factor required to normalise a B-spline basis function,
using its truncated power expansion, such that the basis integrates to 1 over its support.
}
\details{
This function uses \code{\link{bspline_alpha_rule}} to compute the coefficients
\eqn{\alpha_j} for the truncated power expansion, and integrates over the support.
}
\examples{
knots <- c(0, 1, 2, 3, 4, 5)
bspline_normalisation_constant(knots, i = 1, k = 1)

}
