% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logbp}
\alias{logbp}
\title{Offset Logarithm Transform}
\usage{
logbp(x, b = 1e-06, base = 10^1)
}
\arguments{
\item{x}{A numeric vector or matrix of values to transform.}

\item{b}{A small positive offset added to \code{x} before applying the logarithm. Defaults to \code{1e-6}.}

\item{base}{The base of the logarithm. Defaults to \code{10}.}
}
\value{
A numeric vector or matrix with the log-transformed values.
}
\description{
Computes a log-transformation of \code{x} with a small positive bias \code{b},
and an optional logarithmic \code{base}.
}
\examples{
logbp(1:10)
logbp(0, b = 1e-4)
logbp(c(0.1, 1, 10), base = exp(1))

}
