% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittle_optimisation.R
\name{whittle_loglik}
\alias{whittle_loglik}
\title{Whittle Log-Likelihood Objective}
\usage{
whittle_loglik(c, B, periodogram)
}
\arguments{
\item{c}{Coefficient vector for the spectral basis expansion.}

\item{B}{Design matrix (rows = frequencies, cols = basis functions). Easiest to use the output from \code{build_bspline_design_matrix}}

\item{periodogram}{Vector of periodogram values.}
}
\value{
A scalar value: the negative Whittle log-likelihood.
}
\description{
Computes the negative Whittle log-likelihood for a given coefficient vector.
}
